bug: Update DC SQL client to handle long running queries
The key functional change is to implement proper handling of long running queries.
Previously the client would have just returned zero rows if a query ran for more than about 15 seconds, now the client properly polls the status and fetches the full result range.

An additional improvement is that the error message reporting now better surfaces the details from the structured error message. This should help Agents to better tweak queries in case of errors.

As the complexity of the code increases this change also does some refactorings to improve maintainability:
- The DC SQL client is factored into its own filed to separate the MCP related logic from the Client logic. This way it's also more easy to reuse
- The OAuth code is factored out to be reusable by both the server.py as well as the DC SQL client, the global state is replaced by config & session objects
- Logging is added to the whole code base to simplify debugging of issues.